fix: move Development panel read-model scenarios to test_support (#4175)
Relocate scenario pins from development_panel_read_model_test to
colonizethis_economy_test_support so repo.economy_test_no_local_run_scenarios
passes in CI for PR #4203.

Refs #4175
